src/test/java/seedu/address/storage/XmlAdaptedCinemaTest.java
package seedu.address.storage;

import static org.junit.Assert.assertEquals;
import static seedu.address.storage.XmlAdaptedCinema.MISSING_FIELD_MESSAGE_FORMAT;
import static seedu.address.testutil.TypicalCinemas.BENSON;

import java.util.ArrayList;
import java.util.List;
import java.util.stream.Collectors;

import org.junit.Test;

import seedu.address.commons.exceptions.IllegalValueException;
import seedu.address.model.cinema.Address;
import seedu.address.model.cinema.Email;
import seedu.address.model.cinema.Name;
import seedu.address.model.cinema.Phone;
import seedu.address.model.cinema.Theater;
import seedu.address.testutil.Assert;

public class XmlAdaptedCinemaTest {
    private static final String INVALID_NAME = "R@chel";
    private static final String INVALID_PHONE = "+651234";
    private static final String INVALID_ADDRESS = " ";
    private static final String INVALID_EMAIL = "example.com";
    private static final String INVALID_THEATER = "5";
    private static final String INVALID_TAG = "#friend";

    private static final String VALID_NAME = BENSON.getName().toString();
    private static final String VALID_PHONE = BENSON.getPhone().toString();
    private static final String VALID_EMAIL = BENSON.getEmail().toString();
    private static final String VALID_ADDRESS = BENSON.getAddress().toString();
    private static final List<XmlAdaptedTheater> VALID_THEATER = BENSON.getTheaters().stream()
            .map(XmlAdaptedTheater::new)
            .collect(Collectors.toList());
    private static final List<XmlAdaptedTag> VALID_TAGS = BENSON.getTags().stream()
            .map(XmlAdaptedTag::new)
            .collect(Collectors.toList());

    @Test
    public void toModelType_validCinemaDetails_returnsCinema() throws Exception {
        XmlAdaptedCinema cinema = new XmlAdaptedCinema(BENSON);
        assertEquals(BENSON, cinema.toModelType());
    }

    @Test
    public void toModelType_invalidName_throwsIllegalValueException() {
        XmlAdaptedCinema cinema =
                new XmlAdaptedCinema(INVALID_NAME, VALID_PHONE, VALID_EMAIL, VALID_ADDRESS, VALID_THEATER, VALID_TAGS);
        String expectedMessage = Name.MESSAGE_NAME_CONSTRAINTS;
        Assert.assertThrows(IllegalValueException.class, expectedMessage, cinema::toModelType);
    }

    @Test
    public void toModelType_nullName_throwsIllegalValueException() {
        XmlAdaptedCinema cinema = new XmlAdaptedCinema(null, VALID_PHONE, VALID_EMAIL, VALID_ADDRESS, VALID_THEATER, VALID_TAGS);
        String expectedMessage = String.format(MISSING_FIELD_MESSAGE_FORMAT, Name.class.getSimpleName());
        Assert.assertThrows(IllegalValueException.class, expectedMessage, cinema::toModelType);
    }

    @Test
    public void toModelType_invalidPhone_throwsIllegalValueException() {
        XmlAdaptedCinema cinema =
                new XmlAdaptedCinema(VALID_NAME, INVALID_PHONE, VALID_EMAIL, VALID_ADDRESS, VALID_THEATER, VALID_TAGS);
        String expectedMessage = Phone.MESSAGE_PHONE_CONSTRAINTS;
        Assert.assertThrows(IllegalValueException.class, expectedMessage, cinema::toModelType);
    }

    @Test
    public void toModelType_nullPhone_throwsIllegalValueException() {
        XmlAdaptedCinema cinema = new XmlAdaptedCinema(VALID_NAME, null, VALID_EMAIL, VALID_ADDRESS, VALID_THEATER, VALID_TAGS);
        String expectedMessage = String.format(MISSING_FIELD_MESSAGE_FORMAT, Phone.class.getSimpleName());
        Assert.assertThrows(IllegalValueException.class, expectedMessage, cinema::toModelType);
    }

    @Test
    public void toModelType_invalidEmail_throwsIllegalValueException() {
        XmlAdaptedCinema cinema =
                new XmlAdaptedCinema(VALID_NAME, VALID_PHONE, INVALID_EMAIL, VALID_ADDRESS, VALID_THEATER, VALID_TAGS);
        String expectedMessage = Email.MESSAGE_EMAIL_CONSTRAINTS;
        Assert.assertThrows(IllegalValueException.class, expectedMessage, cinema::toModelType);
    }

    @Test
    public void toModelType_nullEmail_throwsIllegalValueException() {
        XmlAdaptedCinema cinema = new XmlAdaptedCinema(VALID_NAME, VALID_PHONE, null, VALID_ADDRESS, VALID_THEATER, VALID_TAGS);
        String expectedMessage = String.format(MISSING_FIELD_MESSAGE_FORMAT, Email.class.getSimpleName());
        Assert.assertThrows(IllegalValueException.class, expectedMessage, cinema::toModelType);
    }

    @Test
    public void toModelType_invalidAddress_throwsIllegalValueException() {
        XmlAdaptedCinema cinema =
                new XmlAdaptedCinema(VALID_NAME, VALID_PHONE, VALID_EMAIL, INVALID_ADDRESS, VALID_THEATER, VALID_TAGS);
        String expectedMessage = Address.MESSAGE_ADDRESS_CONSTRAINTS;
        Assert.assertThrows(IllegalValueException.class, expectedMessage, cinema::toModelType);
    }

    @Test
    public void toModelType_nullAddress_throwsIllegalValueException() {
        XmlAdaptedCinema cinema = new XmlAdaptedCinema(VALID_NAME, VALID_PHONE, VALID_EMAIL, null, VALID_THEATER, VALID_TAGS);
        String expectedMessage = String.format(MISSING_FIELD_MESSAGE_FORMAT, Address.class.getSimpleName());
        Assert.assertThrows(IllegalValueException.class, expectedMessage, cinema::toModelType);
    }
/**
    @Test
    public void toModelType_invalidTheaters_throwsIllegalValueException() {
        List<XmlAdaptedTheater> invalidTheater = new ArrayList<>(VALID_THEATER);
        invalidTheater.add(new XmlAdaptedTheater(INVALID_THEATER));
        XmlAdaptedCinema cinema =
                new XmlAdaptedCinema(VALID_NAME, VALID_PHONE, VALID_EMAIL, VALID_ADDRESS, invalidTheater, VALID_TAGS);
        Assert.assertThrows(IllegalValueException.class, cinema::toModelType);
    }
*/
    @Test
    public void toModelType_invalidTags_throwsIllegalValueException() {
        List<XmlAdaptedTag> invalidTags = new ArrayList<>(VALID_TAGS);
        invalidTags.add(new XmlAdaptedTag(INVALID_TAG));
        XmlAdaptedCinema cinema =
                new XmlAdaptedCinema(VALID_NAME, VALID_PHONE, VALID_EMAIL, VALID_ADDRESS, VALID_THEATER, invalidTags);
        Assert.assertThrows(IllegalValueException.class, cinema::toModelType);
    }

}
